requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Limiter==3.5.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
